Provide sqrt, etc. for `Dimension` and `Magnitude` (#136)
I first tried providing these generically, for all parameter packs. That
turned out to be a bad idea, because you get a lot of false positives
for what a "pack" is.  For example, `Milli<Meters>` can look like a
"pack" of `Meters`!

Then I settled on putting them in a new target, `:power_aliases`, which
is easy to opt into.  All you have to do is implement `pow` and `root`,
and include the file.  I did so for `Dimension` and `Magnitude`, and
added unit tests for each.  Along the way, I also added the missing
arithmetic operators `*` and `/` for `Dimension`.

Includes unit tests and doc updates.

Fixes #84.