Permit integer division with raw numeric divisor
There are a _lot_ of instances of this in the AV repo, mostly in
embedded code.  As I worked through updating them, I realized that
replacing these instances with `integer_quotient()` was making the code
worse.  I think what we're really worried about here is mixing integer
division with unit conversions.  This overload is the only one that
doesn't change the units.

We'll still provide `integer_quotient()` for this case just for
completeness, though.