Update alternatives matrix (#147)
Our alternatives matrix is long overdue for a refresh.

The first core change is to explicitly list each library, indicate which
version we evaluated, and link to it.  We also say a few words about why
that library made our shortlist.  Along this vein, we add a new library,
bernedom/SI, to the list.  It's an impressively popular library (400
GitHub stars!) which didn't get included when we released the library
simply because it wasn't on my radar: at the time, it didn't have the
`dimensional-analysis` tag which most of the others had, so it didn't
show up in my searches.  This fixes #108.

We also clarify for the first time that the rows in our "big" matrix are
ordered (very roughly) by importance, which is a critical piece of
context for the reader.

We add several new rows to our comparison, namely:

- Unit-aware I/O
- Unit-aware math
- Linear algebra
- Physical constants

Next, I updated the results for mp-units for the first time since its
major 2.0 API upgrade.  mp-units improved significantly, and those
improvements were concentrated in the more important categories, such as
"Unit Safety" and "Composability".  (It also regressed in "Macros", but
that's the least important category.)

Finally, I made a few tweaks for previous assessments that I couldn't
really justify in retrospect.