Add CI jobs for Windows (#151)
The job builds a single-file version of the Au library, and runs a
simple test.  This test is enough to reproduce the errors in #143:

https://github.com/aurora-opensource/au/actions/runs/5496101666/job/14882361539

It also catches another few compiler errors on other versions of MSVC,
which we fix here.  The issue is that function parameters can't be used
in a `constexpr` context in general.  These particular parameters are
monovalue types

(https://aurora-opensource.github.io/au/main/reference/detail/monovalue_types/),
so the compiler's being a little pedantic in enforcing this rule, but
this also means that it's easy to fix the error.

With this PR, we go from 0 to 1 in terms of CI jobs for "best effort"
platforms.  This gives us a pattern we can use for future new platforms.

Helps #144.