Add Rep-named aliases for `int`, `unsigned int` (#150)
This is all about the ability to express intent.  Despite their reduced
specificity compared to, e.g., `int32_t`, `unsigned int` and `int` are
very common for end users to write in code.  But right now it's
disproportionately hard for them to spell the corresponding `Quantity`
type concisely.  I'm not sure if we'll add other aliases such as `long`
or `long long`, but these two seem like a clear improvement.

I also updated an obsolete comment which no longer referred to the
correct Rep-named alias.