Finish writing `QuantityPoint` reference
We take the `Quantity` reference as a template, and adjust the contents
accordingly.

Many sections can be omitted.  For example, we don't support
multiplication or division, and there is no "corresponding quantity
point" machinery at this time.

Other concepts are unique to `QuantityPoint`.  For example, addition and
subtraction always involve both `QuantityPoint` and `Quantity`.  Also,
there are further restrictions on implicit conversions in the case of
units with offset origins that can't be represented as integers.

A few other docs got updated as well.

- Fixed a typo in the `Quantity` reference.
- Added a link anchor in the `Unit` reference.
- Added a section on quantity point makers in the unit slots discussion.

To test this change, I clicked on every link and made sure it went where
I was expecting.

I also opened the `Quantity` and `QuantityPoint` reference docs in two
neighboring tabs, and clicked on every corresponding section one at a
time, tabbing back and forth to see the changes.

Fixes #140.