Clarify "dangerous conversion" error messages (#158)
The status quo has two major problems.  First, it gives out-of-context
advice to use the explicit-Rep version `.as<Rep>`, which (as the
troubleshooting guide explains) is usually not the _best_ remedy.
Second, it gives the user no hint that the true root cause is likely the
use of integer Rep, which is the only case where this message can occur.

The remedy has two parts.  First, we make that assumption --- that this
never occurs unless we have an integral Rep --- explicit, testing it
with a new `static_assert` (one that we _never_ expect end users to
see).  This motivated encapsulating the conditions into named variables,
which incidentally improves the compiler error messages.  Second,
instead of telling users to "just use `.as<Rep>()`", we link directly to
the relevant section of the troubleshooting guide, so they can see what
remedies are available and understand their tradeoffs.

This change, of course, makes the troubleshooting guide out-of-date.
Instead of updating only that section, I regenerated all of it by
re-running the compiler on the example commands.  There are several
categories of changes.

1. **Line numbers.**  By far the most numerous change.  Both the Au
   library files, and the `error_examples.cc` file, have changed (the
   latter, via inclusion of the full copyright header).

2. **Dangerous conversion.**  This, of course, stems from the main
   change in this PR.

3. **Integer division.**  We now encapsulate this logic inside a helper,
   `warn_if_integer_division`, which changes the error message.

4. **Dangerous inversion.**  Instead of a raw literal `1000`, we now
   have a variable called `threshold` (and the _value_ of that variable
   has also changed).

Fixes #127.