Make Au compatible with `-Wconversion`
The goal is for Au to not raise any warnings for users who build with
`-Wconversion`.  This is subtly different than the goal of enabling
`-Wconversion` on our own builds.  Some of our test cases would fail
this warning, but I still feel they are important for giving us
confidence that the library is correct.

We can test this change with the following command:

```sh
bazel test \
  --copt=-Wconversion \
  --copt=-Werror \
  //au/...:all
```

After this PR, everything passes except for the quantity test cases
mentioned above.