Implement `clamp(v, lo, hi)` (#165)
This is similar to `std::clamp(v, lo, hi)`, introduced in C++17, but
with some interesting twists due to handling units.  Similar to `min`
and `max`, we express the return value in the common unit of all inputs.
Once again, this also means we return by value instead of reference.

One consequence, which may be counterintuitive at first, is that the
return type can be different from the type of the value we're clamping!
However, this makes sense after more careful consideration.  If we ask
to clamp a value in, say, _meters_, to a range whose bounds we express
in _millimeters_, then if the value falls outside this range, _of
course_ we'll need to express it in millimeters.  Since the return type
can't depend on whether the value is outside the range, this is
sufficient.

(Of course, I expect the usual case when units are different is that the
value is already more precise than the bounds.  And I expect the usual
_usual_ case is that the units are _not_ different.)

`QuantityPoint` is even more interesting.  It turns out we need to take
the origin offset into account whenever it exists.  Imagine clamping 0
degrees Celsius to within two values in Kelvins.  We can't compare the
value to the bounds without adding the relative offset, which can't be
expressed more coarsely than 1/20 Kelvins.  Fortunately, our "common
point unit" machinery handles this case with ease.

Includes unit tests for both `Quantity` and `QuantityPoint`, as well as
reference documentation.

Fixes #42.