Add style guide (#164)
I skimmed through the Google C++ style guide, and looked for any
differences in how we do things.  I called out the ones I found.  We can
add more here as we encounter them.

The result can be previewed here:

https://github.com/aurora-opensource/au/blob/style-guide%2322/CONTRIBUTING.md#style-guide

To test, I clicked every link.  They all go to the right anchor,
although some of the later ones seemed to be a little "off" to me in
their location.  However, the link is definitely correct, so I think
this is a quirk we can't do anything about.

Fixes #22.

---------

Co-authored-by: johnzoppina <zoppina@aurora.tech>