Document migration facilities (#162)
This PR adds how-to guides for the migration facilities.  We include a
general how-to guide, and a specific guide for the built-in migration
facilities for the nholthaus library.  We also add some more link
anchors in other docs as needed, and remove the howto stub from the
`corresponding_quantity.md` reference page.

Fixes #107.

---------

Co-authored-by: johnzoppina <zoppina@aurora.tech>