Add --all-units flag to single file script
This flag will cause the script to act as if the `--units` flag had been
populated with every unit that is known to the library.

Under the hood, the script has a safety mechanism to make sure that we
actually use every command line parameter in the script.  This new
parameter would run afoul of this mechanism, because it's mutually
exclusive with the `--units` argument.  Therefore, we use `--all-units`
to populate `--units`, and then delete `--all-units`.