Support `Zero` in `min`, `max`, and `clamp` (#182)
We focus on the use cases that provide the clearest value: that is,
either argument for `min` and `max`, and either boundary argument for
`clamp`.

There are other use cases --- for example, both arguments `Zero` for
`max`, or the 6 remaining combinatoric possibilities for `clamp`.  But
we don't plan to support them because they don't seem very useful.

I don't think this needs a doc update, because the point of this PR is
to make these APIs act the way that users probably already expect.

Fixes #179.