Update release instructions (#180)
We now include running the commented-out test cases.  This closes a gap
in our testing.  (It's not as good as running them automatically on
every commit, but we don't have an easy way to do that.)

We also make sure to check the cumulative compile time impact, to make
sure no significant regressions have slipped in without our notice.