Add unit-aware overload for `std::remainder`
This is basically like `std::fmod`, but it's zero-centered.  It's really
handy for things like rectifying angles when we'd rather express
`degrees(270)` as `degrees(-90)`.

To make this change, I grepped the codebase for `fmod`, and simply made
whatever change was reasonable.  One of these changes was to factor out
a helper to make it more generic; I did this in a separate "checkpoint
commit" on the PR to make it easier to review.