Update links to point to main doc website
Several months ago, we versioned our documentation using `mike`.  This
means that any legacy links that don't provide a specific version will
be redirected to the `0.3.2` version of the documentation website.  (Try
this out on our main README page with the alternatives comparison, for
example.)  Only the main landing page is an exception:
(https://aurora-opensource.github.io/au) automatically redirects to
(https://aurora-opensource.github.io/au/main/), so we don't need to
update those.

This PR goes through and updates every link to a non-main page of our
doc website, so that it points directly to the `main` version of the
docs.