Add unit symbols
First, we define the mechanisms, using the new `SymbolFor` monovalue
type, and the `symbol_for` utility function.  We are starting out
conservative here, and only supplying operations for raw numbers,
quantity, and self-composition for now.  It's easy to add more later if
we want; easier than subtracting something we regret.

Next, we define a symbol instance for every existing unit.  Usually,
this is just the unit label.  Exceptions:

- For `Percent`, we use `pct` instead of `%` (which would not work).
- For temperatures with "offset zero", we append `_qty` as an extra
  measure to avoid confusion: `degC_qty` and `degF_qty`.  (If users
  don't like this, they can define their own symbols!)

Many symbols would be in danger of colliding with, e.g.,
`SingularNameFor`.  For example, `bar` is both the symbol for `Bars` and
its singular name.  For this reason --- and, because symbols introduce
many very short names! --- we add an extra namespace, `::au::symbols`.
Users should import symbols one at a time, as needed.  For example, in a
`.cc` file:

```cpp
using ::au::symbols::m;
```

Finally, we update the docs.  For now, we just provide reference docs,
update the how-to, and update the comparison chart.  Since unit symbols
are simply better than UDLs, we bump the UDL libraries down from "good"
to "fair". mp-units gets bumped down from "best" to "good" because now
there is no "best".