Add unit symbol for every unit
Usually, this is just the unit label. Exceptions:

- For Percent, we use `pct` instead of `%` (which would not work).
- For temperatures with "offset zero", we append `_qty` as an extra
  measure to avoid confusion: `degC_qty` and `degF_qty`.  (If users
  don't like this, they can define their own symbols!)

Many symbols would be in danger of colliding with, e.g.,
`SingularNameFor`. For example, `bar` is both the symbol for `Bars` and
its singular name. For this reason --- and, because symbols introduce
many very short names! --- we add an extra namespace, `::au::symbols`.
Users should import symbols one at a time, as needed. For example, in a
`.cc` file:

```cpp
using ::au::symbols::m;
```

This makes it easy for anyone who runs across the `m` symbol to find its
definition.

Fixes #43.