Add unit symbols (#196)
This PR defines the mechanisms for symbols, using the new `SymbolFor`
monovalue type, and the `symbol_for` utility function.  We are starting
out conservative here, and only supplying operations for raw numbers,
quantity, and self-composition for now.  It's easy to add more later if
we want; easier than subtracting something we regret.

`SymbolFor` supports prefixes, too.  We can apply a prefix applier to an
instance, which makes it easy to create a symbol called `nm` as
`nano(m)` if we already have a symbol `m`.

We also update the docs.  For now, we just provide reference docs,
update the how-to, and update the comparison chart.  Since unit symbols
are simply better than UDLs, we bump the UDL libraries down from "good"
to "fair". mp-units gets bumped down from "best" to "good" because now
there is no "best".

Helps #43.