Add discussion of overflow and how to mitigate it
This finally --- _finally!_ --- gives us an authoritative place to link
for explaining the "overflow safety surface".  I was going to make that
the topic of the page, but I realized that there's a lot more value in
discussing the overflow problem more generally!  I expect this link will
be a useful reference for other units libraries as well.

I adapted some contents that were hidden away a couple layers deep in
the 103 tutorial, and replaced those contents with a link to the new
page.

This helps #90 tangentially, in that I wanted to link to this page in
the docs for `Constant`, but was stymied by its nonexistence.