Use perfect conversion policy in inverse functions
Now that we have constants, we can replace the overflow safety surface
heuristic with exact checks.  This PR uses constants to do this.  It
adds a new unit test that would have failed before these implementation
changes.  It also includes an uncomment-to-check test case which still
would fail, as it must.

Fixes #190.